+++ app/.mocharc.js
@@ -1,5 +1,8 @@
 "use strict";
 
+/**
+ * @type {import('@mocha').AppConfig}
+ */
 module.exports = {
   recursive: false,
   spec: ["app/**/*.spec.ts"], // the positional arguments!