+++ app/types/types.d.ts
@@ -1,0 +1,5 @@
+import { Opaque } from "type-fest";
+
+declare global {
+  type UnixTimestamp = Opaque<number, "UnixTimestamp">;
+}